Make datum subclasses share more via parent class
