+++ backend/src/org/commcare/suite/model/SessionDatum.java
@@ -1,26 +1,44 @@
 package org.commcare.suite.model;
 
-import org.javarosa.core.model.instance.TreeReference;
+import org.javarosa.core.util.externalizable.DeserializationException;
+import org.javarosa.core.util.externalizable.ExtUtil;
 import org.javarosa.core.util.externalizable.Externalizable;
+import org.javarosa.core.util.externalizable.PrototypeFactory;
+
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
 
 /**
  * @author ctsims
  */
-public abstract class SessionDatum implements Externalizable{
+public abstract class SessionDatum implements Externalizable {
+    private String value;
+    private String id;
+
+    protected SessionDatum(String id, String value) {
+        this.id = id;
+        this.value = value;
+    }
 
     public String getDataId() {
         return id;
     }
 
-    /**
-     * the ID of a detail that will show a selected item for confirmation. If not present,
-     * no confirmation screen is shown after item selection
-     */
-    public String getLongDetail() {
-        return longDetail;
-    }
-
     public String getValue() {
         return value;
+    }
+
+    @Override
+    public void readExternal(DataInputStream in, PrototypeFactory pf)
+            throws IOException, DeserializationException {
+        id = ExtUtil.nullIfEmpty(ExtUtil.readString(in));
+        value = ExtUtil.readString(in);
+    }
+
+    @Override
+    public void writeExternal(DataOutputStream out) throws IOException {
+        ExtUtil.writeString(out, ExtUtil.emptyIfNull(id));
+        ExtUtil.writeString(out, value);
     }
 }
